moved SimpleAgent to ML, added headers with ideas for learning process
