+++ src/Makefile
@@ -2,12 +2,12 @@ all:
 	cd SequenceSimulation; make all
 	cd visualization; make all
 	cd alignment; make all
-	cd ReinforcementLearning/SimpleAgent; make all
+	cd ML/SimpleAgent; make all
 	cd GameMain; make all
 
 clean:
 	cd SequenceSimulation; make clean
 	cd visualization; make clean
 	cd alignment; make clean
-	cd ReinforcementLearning/SimpleAgent; make clean
+	cd ML/SimpleAgent; make clean
 	cd GameMain; make clean
